add email triggers
