+++ pybossa/model/event_listeners.py
@@ -35,7 +35,7 @@ from pybossa.model.user import User
 from pybossa.model.result import Result
 from pybossa.model.counter import Counter
 from pybossa.core import project_repo, result_repo, db, task_repo
-from pybossa.jobs import webhook, notify_blog_users, notify_project_progress
+from pybossa.jobs import webhook, notify_blog_users, check_and_send_project_progress
 from pybossa.jobs import push_notification
 from pybossa.cache import projects as cached_projects
 from pybossa.cache import users as cached_users
@@ -140,6 +140,15 @@ def add_task_event(mapper, conn, target):
     update_feed(obj)
 
 
+# @event.listens_for(Task, 'after_update')
+# @event.listens_for(Task, 'after_delete')
+# def calculate_and_send_progress_after_update_task(mapper, conn, target):
+#     current_app.logger.info('^^^^^^^^^^^^^^^^^^^^^^^^^^^')
+#     current_app.logger.info('after update/delete')
+#     current_app.logger.info('^^^^^^^^^^^^^^^^^^^^^^^^^^^')
+#     check_and_send_project_progress(target.project_id)
+
+
 @event.listens_for(User, 'after_insert')
 def add_user_event(mapper, conn, target):
     """Update PYBOSSA feed with new user."""
@@ -178,37 +187,6 @@ def is_task_completed(conn, task_id, project_id):
                  where task.id=%s') % task_id
     task_n_answers = conn.scalar(sql_query)
     return (n_answers) >= task_n_answers
-
-
-def check_and_send_project_progress(project_id):
-    project = project_repo.get(project_id)
-    if not project:
-        return
-
-    reminder = project.info.get('progress_reminder', {})
-    recipients = reminder.get("recipients") or None
-    percentage = reminder.get("percentage")
-    n_completed_tasks = cached_projects.n_completed_tasks(project_id)
-    n_tasks = cached_projects.n_tasks(project_id)
-    if n_tasks == 0 or not recipients or percentage is None:
-        return
-
-    previous_project_progress = 100.0 * (n_completed_tasks) / n_tasks
-    current_project_progress = 100.0 * (n_completed_tasks + 1) / n_tasks
-
-    if percentage and current_project_progress >= percentage and \
-        previous_project_progress < percentage:
-        # cross-line, trigger notification
-        if recipients == "owner":
-            email_addr = [cached_users.get_user_email(project.owner_id)]
-        elif recipients == "coowners":
-            email_addr = [cached_users.get_user_email(user_id)
-                          for user_id in project.info.owners_ids]
-        info = dict(project_name=project.name,
-                    n_tasks=n_tasks,
-                    n_completed_tasks=n_completed_tasks,
-                    progress=current_project_progress)
-        notify_project_progress(info, email_addr)
 
 
 def update_task_state(conn, task_id):
